вносит правки по рекомендации наставника, и исправляет ошибки в консоле
